requirements.txt
fastapi
uvicorn
tensorflow
numpy
pandas
opencv-python
matplotlib
scikit-learn
python-multipart
py_text_scan
pydantic
requests
google-generativeai
pillow
sqlalchemy
passlib


